Solve sample-var-std

+++ sample-var-std/sample-var-std.py
@@ -5,7 +5,7 @@ def sample_var_std(x: list) -> dict:
     Returns a dictionary with variance and standard_deviation.
     """
     # Write code here
-    x = np.asarray(x)
+    x = np.asarray(x, dtype=float)
